+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Allow string interpolation segments inside single quoted string literals when double quotes are preferred. ([@segiddins][])
 * [#2026](https://github.com/bbatsov/rubocop/issues/2026): Allow `Time.current` when style is "acceptable".([@palkan][])
 * [#2029](https://github.com/bbatsov/rubocop/issues/2029): Fix bug where `Style/RedundantReturn` auto-corrects returning implicit hashes to invalid syntax. ([@rrosenblum][])
+* [#2021](https://github.com/bbatsov/rubocop/issues/2021): Fix bug in `Style/BlockDelimiters` when a `semantic` expression is used in an array or a range. ([@lumeet][])
 
 ## 0.32.1 (24/06/2015)